ipdata-geolocation-api-microservice/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
redis
elasticsearch
aiohttp
backoff
datetime
pydantic
uuid
python-dateutil
pandas
